refactor: improve encapsulation of ui connection
